Change exit codes for alerts and quick-scan commands
Change exit codes for the alerts and quick-scan commands to more closely
match the behaviour of ZAP's scan scripts provided in the Docker image.
Now if any alerts are found for the given level, the command will exit with
a status of 1, and any other errors will produce the status code 2.